+++ orb/forms.py
@@ -235,72 +235,10 @@ class HeaderSearchForm(forms.Form):
                     FieldWithButtons('q',Submit('submit', _(u'Search'), css_class='btn btn-default')),
                     Row (HTML('<a href="' + reverse('orb_search_advanced') + '">Advanced search</a>'), css_class='advanced-search-link'),
                     )
-        
-class TagFilterForm(forms.Form):
-    health_topic = forms.MultipleChoiceField(
-                        widget=forms.CheckboxSelectMultiple,
-                        required=False,)
-    resource_type = forms.MultipleChoiceField(
-                        widget=forms.CheckboxSelectMultiple,
-                        required=False,)
-    audience = forms.MultipleChoiceField(
-                        widget=forms.CheckboxSelectMultiple,
-                        required=False,)
-    geography = forms.MultipleChoiceField(
-                        widget=forms.CheckboxSelectMultiple,
-                        required=False,)
-    language = forms.MultipleChoiceField(
-                        widget=forms.CheckboxSelectMultiple,
-                        required=False,)
-    device = forms.MultipleChoiceField(
-                        widget=forms.CheckboxSelectMultiple,
-                        required=False,)
-    license = forms.MultipleChoiceField(
-                        widget=forms.CheckboxSelectMultiple,
-                        required=False,)
-    
-    def __init__(self, *args, **kwargs):
-        super(TagFilterForm, self).__init__(*args, **kwargs)
-        self.helper = FormHelper()
-        self.helper.form_class = 'form-horizontal'
-        self.helper.label_class = 'col-lg-2'
-        self.helper.field_class = 'col-lg-8'
-        self.helper.layout = Layout(
-                'health_topic',
-                Row (HTML('<hr>')),
-                'resource_type',
-                Row (HTML('<hr>')),
-                'audience',
-                Row (HTML('<hr>')),
-                'geography',
-                Row (HTML('<hr>')),
-                'language',
-                Row (HTML('<hr>')),
-                'device',
-                Row (HTML('<hr>')),
-                'license',
-                Row (HTML('<hr>')),
-                Div(
-                   Submit('submit', _(u'Go'), css_class='btn btn-default'),
-                   css_class='col-lg-offset-2 col-lg-8',
-                ),
-            )
-    def clean(self):
-        empty = True
-        
-        for name,slug in settings.TAG_FILTER_CATEGORIES:
-            tag_ids = self.cleaned_data.get(name)
-            if tag_ids:
-                empty = False
-        
-        if empty:
-            raise forms.ValidationError( _(u"Please select at least one tag to filter on"))
-            
-        return self.cleaned_data
 
 class AdvancedSearchForm(forms.Form):
     q = forms.CharField(
-                required=True,
+                required=False,
                 label= _(u'Search terms'),
                 error_messages={'required': _(u'Please enter something to search for')},)
     health_topic = forms.MultipleChoiceField(
@@ -354,6 +292,20 @@ class AdvancedSearchForm(forms.Form):
                 ),
             )
 
+    def clean(self):
+        empty_tags = True
+        query = self.cleaned_data.get("q","").strip()
+        
+        for name,slug in settings.ADVANCED_SEARCH_CATEGORIES:
+            tag_ids = self.cleaned_data.get(name)
+            if tag_ids:
+                empty_tags = False
+        
+        if empty_tags and query == "":
+            raise forms.ValidationError( _(u"Please select at least a search term or a tag"))
+            
+        return self.cleaned_data
+    
 class ResourceRejectForm(forms.Form):
     criteria = forms.MultipleChoiceField(
                         widget=forms.CheckboxSelectMultiple,