+++ orb/local_settings.py
@@ -28,7 +28,7 @@ def modify(settings):
     }                                  
     settings['HAYSTACK_SIGNAL_PROCESSOR'] = 'haystack.signals.RealtimeSignalProcessor'
     
-    settings['TAG_FILTER_CATEGORIES'] = [('health_topic','health-domain' ),
+    settings['ADVANCED_SEARCH_CATEGORIES'] = [('health_topic','health-domain' ),
                                          ('resource_type','type'), 
                                          ('audience', 'audience'), 
                                          ('geography', 'geography'),